Refactor SelectionOverlay into separate file
